+++ src/config.ts
@@ -1,4 +1,4 @@
 export const APP_NAME = 'EA Media Tools';
-export const APP_UPDATE_REPOSITORY = 'eaforlife/EA-Media-Tools';
+export const APP_UPDATE_REPOSITORY = 'eaforlife/media-converter';
 export const FFMPEG_RELEASE_API =
   'https://api.github.com/repos/jellyfin/jellyfin-ffmpeg/releases/latest';